requirements.txt
python-binance
Ta-lib
numpy
flask
backtrader(?)